% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subnetsimInternalMethods.R
\name{getSubNet}
\alias{getSubNet}
\title{Counting the number of links in a network.}
\usage{
getSubNet(netAll, nodesSel)
}
\arguments{
\item{netAll}{a \code{hash} representing the global network.}

\item{nodesSel}{a \code{vector} of \code{caracter} representing all the
nodes contained in the sub-network.}
}
\value{
a \code{integer} representing the number of links.
}
\description{
Counting the number of links in a network.
}
\examples{

## TODO

}
\author{
Pascal Belleau, Astrid Deschenes
}
\keyword{internal}
