% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkInternalMethods.R
\name{networkFromSif}
\alias{networkFromSif}
\title{Generate a network object.}
\usage{
networkFromSif(netFileName, subNetFileName = NULL)
}
\arguments{
\item{netFileName}{TODO}

\item{subNetFileName}{TODO}
}
\value{
a \code{data.frame} containing all simulated sub-network.
}
\description{
Generation of a network list from sif.
}
\examples{

networkFile <- system.file("extdata", "demo_network.sif",
package="subnetsim")

## Create a network without the subnetwork information
demo_network <- networkFromSif(netFileName = networkFile)

subnetworkFile <- system.file("extdata", "demo_subnetwork.sif",
package="subnetsim")

## Create a network with the subnetwork information
demo_network_with_sub <- networkFromSif(netFileName = networkFile,
    subNetFileName = subnetworkFile)

}
\author{
Pascal Belleau, Astrid Deschenes
}
\keyword{internal}
