% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subnetsim.R
\docType{package}
\name{subnetsim-package}
\alias{subnetsim-package}
\alias{subnetsim}
\title{subnetsim: Permutation-Based Analysis}
\description{
This package does a permutation analysis, based on Monte Carlo sampling,
for testing the hypothesis that the number of observed linked inside
a subnetwork is significantly higher that the number obtained by a
stochastic effect.
}
\author{
Pascal Belleau

Maintainer:
Pascal Belleau <pascal_belleau@hotmail.com>
}
\keyword{package}
