% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subnetwork.R
\name{subnetwork}
\alias{subnetwork}
\title{TODO}
\usage{
subnetwork(network, nbIter = 10000, nbNodes = NULL, seedV = -1)
}
\arguments{
\item{network}{an object \code{network} TODO}

\item{nbIter}{a single \code{numeric}, interpreted as an \code{integer}
indicating the number of iterations to be run. Default: \code{10000}.}

\item{nbNodes}{a single \code{numeric}, interpreted as an \code{integer}
TODO.}

\item{seedV}{a single \code{numeric}, interpreted as an \code{integer}, a
seed used when reproducible results are
needed. When a value inferior or equal to zero is given, a random integer
is used. Default: \code{-1}.}
}
\value{
a list is returned for use by \code{print.subnetwork}. This list has
           the components:
           \itemize{
               \item{ TODO }
               \item{ TODO }
           }
}
\description{
TODO
}
\examples{
## TODO


}
\author{
Pascal Belleau
}
\keyword{subnetwork}
