% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subnetsimInternalMethods.R
\name{getOneLink}
\alias{getOneLink}
\title{Counting the number of nodes and the number of links in a
network.}
\usage{
getOneLink(netAll, nodesSel)
}
\arguments{
\item{netAll}{TODO}

\item{nodesSel}{TODO}
}
\value{
a \code{vector} containing 2 \code{integer}. The first \code{integer}
represent the number of nodes and the second \code{integer} represents the
number of links.
}
\description{
Counting the number of nodes and the number of links in a
network.
}
\examples{

## TODO

}
\author{
Pascal Belleau, Astrid Deschenes
}
\keyword{internal}
