% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subnetsimInternalMethods.R
\name{getOneLink}
\alias{getOneLink}
\title{Counting the number of nodes and the number of links associated to a
subnetwork.}
\usage{
getOneLink(netAll, nodesSel)
}
\arguments{
\item{netAll}{a \code{hash} representing the global network.}

\item{nodesSel}{a \code{vector} of \code{caracter} strings representing all the
nodes contained in the subnetwork.}
}
\value{
a \code{vector} containing 2 \code{integer}. The first
\code{integer} represents the number of nodes and the second \code{integer}
represents the number of links associated to the tested subnetwork.
}
\description{
Counting the number of nodes and the number of links in a
specific subnetwork. From a list of nodes forming the subnetwork, the number of
nodes and the number of links associated to the subnetwork are calculated.
The nodes and links inside the subnetwork are counted as well as all first-degree
neighbor nodes and all links associated to the first-degree nodes.
}
\examples{

## TODO

}
\author{
Pascal Belleau, Astrid Deschenes
}
\keyword{internal}
