% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subnetsimInternalMethods.R
\name{getSubNet}
\alias{getSubNet}
\title{Counting the number of links in a network.}
\usage{
getSubNet(netAll, nodesSel)
}
\arguments{
\item{netAll}{TODO}

\item{nodesSel}{TODO}
}
\value{
a \code{integer} representing the number of lilnks.
}
\description{
Counting the number of links in a network.
}
\examples{

## TODO

}
\author{
Pascal Belleau, Astrid Deschenes
}
\keyword{internal}
