% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{network}
\alias{network}
\title{Generate a network object from a fileType.}
\usage{
network(netFileName, fileType = "sif", subNetFileName = NULL)
}
\arguments{
\item{netFileName}{TODO}

\item{fileType}{TODO}

\item{subNetFileName}{TODO}
}
\value{
a \code{data.frame}
}
\description{
Generation of a network object.
}
\examples{

## TODO

}
\author{
Pascal Belleau, Astrid Deschenes
}
\keyword{internal}
