% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkInternalMethods.R
\name{networkFromSif}
\alias{networkFromSif}
\title{Generate a network object.}
\usage{
networkFromSif(netFileName, subNetFileName = NULL)
}
\arguments{
\item{netFileName}{a \code{character} string for the name of
file in SIF format that contains the information of the
global network.}

\item{subNetFileName}{a \code{character} string for the name of
file in SIF format that contains the information of the
subnetwork. This parameter is not mandatory. The information for
the subnetwork is only extracted when a filename is given.}
}
\value{
a \code{list} containing the following components:
           \itemize{
           \item{ \code{netAll} a \code{hash} containing all
           the information from the global network.}
           \item{ \code{nodesAll} a vector of string representing
           all the nodes present in the global network.}
           \item{ \code{nodesSubNet} a vector of string representing
           all the nodes present in the subnetwork to be tested or
           \code{NULL} when information not known.}
           }
}
\description{
Generation of a list that contain all the
global network information extracted from a sif file. When
a subnetwork file is present, the information from
the subnetwork is added to the list.
}
\examples{

## Access SIF file for global network stored in subnetsim package
networkFile <- system.file("extdata", "demo_network.sif",
package="subnetsim")

## Create a network without the subnetwork information
demo_network <- subnetsim:::networkFromSif(netFileName = networkFile)

## Access SIF file for subnetwork stored in subnetsim package
subnetworkFile <- system.file("extdata", "demo_subnetwork.sif",
package="subnetsim")

## Create a network with the subnetwork information
demo_network_with_sub <- subnetsim:::networkFromSif(netFileName = networkFile,
    subNetFileName = subnetworkFile)

}
\author{
Pascal Belleau, Astrid Deschenes
}
\keyword{internal}
