% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkInternalMethods.R
\name{networkFromSif}
\alias{networkFromSif}
\title{Generate a network object.}
\usage{
networkFromSif(netFileName, subNetFileName = NULL)
}
\arguments{
\item{netFileName}{TODO}

\item{subNetFileName}{TODO}
}
\value{
a \code{data.frame} containing all simulated sub-network.
}
\description{
Generation of a network list from sif.
}
\examples{

## TODO

}
\author{
Pascal Belleau, Astrid Deschenes
}
\keyword{internal}
