% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subnetsimInternalMethods.R
\name{simuleSubNet}
\alias{simuleSubNet}
\title{Generating simulated sub-networks.}
\usage{
simuleSubNet(netAll, nodesAll, nbIter, nbNodes, seedV)
}
\arguments{
\item{netAll}{a \code{hash} representing the global network.}

\item{nodesAll}{\code{list} of all the node in the network}

\item{nbIter}{a \code{integer} indicating the number of iterations to be
run.}

\item{nbNodes}{a \code{integer} indicating the number of nodes that each
simulated sub-network will contain.}

\item{seedV}{a \code{integer}, a seed used when reproducible results are
needed. When a value inferior or equal to zero is given, a random integer
is used.}
}
\value{
a \code{data.frame} containing all simulated sub-network. Each
row of the \code{data.frame} corresponds to one simulation.:
\itemize{
\item nbLink: Number of link starting from each node of the simulate subnetwork
\item nbNodesOneLink: Number of node of the simulate subnetwork
 plus the number of node connected with a node in the subnetwork
\item nbLinkOneLink Number of link starting from each node of the simulate
subnetwork plus the number of link starting from node connected with
a node in the subnetwork
}
}
\description{
Generationg multiple simulated sub-networks from a network. The
sub-networks all have the same number of nodes.
}
\examples{

## TODO

}
\author{
Pascal Belleau, Astrid Deschenes
}
\keyword{internal}
