% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subnetsim.R
\docType{package}
\name{subnetsim-package}
\alias{subnetsim-package}
\alias{subnetsim}
\title{subnetsim: Validating that a specific subnetwork has significantly more
interactions than expected by randomness}
\description{
This package implements a simulation analysis, based on bootstrapping, for
testing the hypothesis that a selected subnetwork, which is part of a larger
network, has more interactions than expected by randomness. To do so,
subnetworks with the same number of nodes than the tested subnetwork are randomly
generated by bootstrapping. Three measures of connectivity are used to validate the
hypothesis: the number of nodes in the subnetwork including all the first-degree
neighbor nodes, the number of links in the subnetwork including all the links
of the first-degree neighbor nodes and the number of links in the subnetwork.
}
\author{
Pascal Belleau, Astrid Deschenes, Martial Boutchueng-Djidjou, Robert L. Faure

Maintainer:
Pascal Belleau <pascal_belleau@hotmail.com>
}
\keyword{package}
