% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subnetsim.R
\docType{package}
\name{subnetsim-package}
\alias{subnetsim-package}
\alias{subnetsim}
\title{subnetsim: Permutation-Based Analysis}
\description{
This package does a permutation analysis, based on Monte Carlo sampling,
for testing the hypothesis that the number of observed linked inside
a subnetwork is significantly higher that the number obtained by a
stochastic effect with a subnetwork with the same number of node.
}
\details{
This package tests if a subnetwork is more connected than a randomly
selected subnetwork with the same number of nodes. We validate three
measures of connectivity: the number of links starting from the nodes
of the subnetwork, the number of nodes plus the number of nodes
directly connected to the subnetwork and the number of links from nodes
of the subnetwork and the nodes directly connected to the subnetwork.
To test if the subnetwork is significantly more connected than connections
obtained by randomness, the package does a simulation analysis.
}
\author{
Pascal Belleau, Astrid Deschenes

Maintainer:
Pascal Belleau <pascal_belleau@hotmail.com>
}
\keyword{package}
