% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.subnetwork.R, R/print.subnetwork.R,
%   R/subnetwork.R
\name{plot.subnetwork}
\alias{plot.subnetwork}
\alias{print.subnetwork}
\alias{subnetwork}
\title{TODO}
\usage{
\method{plot}{subnetwork}(x, bins = 30, ...)

\method{print}{subnetwork}(x, digits = 4, ...)

subnetwork(network, nbIter = 10000, nbNodes = NULL, seedV = -1)
}
\arguments{
\item{x}{the output object from \code{subnetwork} function to be plotted}

\item{bins}{the number of bins used to generate the histograms}

\item{digits}{the number of digits to use when printing numbers}

\item{network}{an object \code{network} TODO}

\item{nbIter}{a single \code{numeric}, interpreted as an \code{integer}
indicating the number of iterations to be run. Default: \code{10000}.}

\item{nbNodes}{a single \code{numeric}, interpreted as an \code{integer}
TODO.}

\item{seedV}{a single \code{numeric}, interpreted as an \code{integer}, a
seed used when reproducible results are
needed. When a value inferior or equal to zero is given, a random integer
is used. Default: \code{-1}.}

\item{\ldots}{arguments passed to or from other methods}

\item{x}{the output object from \code{subnetwork} function to be printed}

\item{\ldots}{arguments passed to or from other methods}
}
\value{
a list is returned for use by \code{print.subnetwork}. This list has
           the following components:
           \itemize{
               \item{ TODO }
               \item{ TODO }
           }
}
\description{
TODO
}
\examples{
## TODO


}
\author{
Pascal Belleau, Astrid Deschenes
}
\keyword{subnetwork}
