% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.subnetwork.R, R/print.subnetwork.R,
%   R/subnetwork.R
\name{plot.subnetwork}
\alias{plot.subnetwork}
\alias{print.subnetwork}
\alias{subnetwork}
\title{Subnetwork simulation through randomness sampling}
\usage{
\method{plot}{subnetwork}(x, bins = 30, ...)

\method{print}{subnetwork}(x, digits = 4, ...)

subnetwork(network, nbIter = 10000, nbNodes = NULL, nbLink = NULL,
  nbNodesOneLink = NULL, nbLinkOneLink = NULL, seedV = -1)
}
\arguments{
\item{x}{the output object from \code{subnetwork} function}

\item{bins}{the number of bins used to generate the histograms}

\item{digits}{the number of digits to use when printing numbers}

\item{network}{an object of \code{class} \code{network} that represents
the global network from which the tested subnetwork is extracted. The
subnetwork can be included as an element of the \code{network} object.
If the subnetwork is included in the \code{network} object, the
\code{nbNodes}, \code{nbLink}, \code{nbNodesOneLink} and
\code{nbLinkOneLink} parameters should all be \code{NULL}. If the
subnetwork is not included, all four parameters should be set.}

\item{nbIter}{a single \code{numeric}, interpreted as an \code{integer}
indicating the number of iterations to be run. Default: \code{10000}.}

\item{nbNodes}{a single \code{numeric}, interpreted as an \code{integer}
the number of nodes of the tested subnetwork. If the value is \code{NULL},
the value is obtained through the subnetwork present in the network object.
Default: \code{NULL}.}

\item{nbLink}{a single \code{numeric}, interpreted as an \code{interger},
the number of links of the tested subnetwork. If the value is \code{NULL},
the value is obtained through the subnetwork present in the network object.
Default: \code{NULL}.}

\item{nbNodesOneLink}{a single \code{numeric}, interpreted as an \code{interger},
the number of nodes of the tested subnetwork including the first-degree neighbor
nodes. If the value is \code{NULL},
the value is obtained through the subnetwork present in the network object.
Default: \code{NULL}.}

\item{nbLinkOneLink}{a single \code{numeric}, interpreted as an \code{interger},
the number of links of the tested subnetwork including all the links of the
first-degree neighbor nodes. If the value is \code{NULL},
the value is obtained through the subnetwork present in the network object.
Default: \code{NULL}.}

\item{seedV}{a single \code{numeric}, interpreted as an \code{integer}, a
seed used when reproducible results are
needed. When a value inferior or equal to zero is given, a random integer
is used. Default: \code{-1}.}

\item{\ldots}{arguments passed to or from other methods}
}
\value{
a list is returned for use by \code{print.subnetwork} and
\code{plot.subnetwork}. This list has the following components:
           \itemize{
               \item{ \code{nbNodesOneLink} a vector of the number of nodes
               for each simulated subnetwork extended to contain the
               first-degree nodes. }
               \item{ \code{nbLinkOneLink} a vector of the number of links
               for each simulated subnetwork extended to contain all
               the links from thefirst-degree nodes. }
               \item{ \code{nbLink} a vector of the number of links
               for each simulated subnetwork. }
               \item{ \code{nbIter} a single \code{integer} corresponding
               to the number of iterations done. }
               \item{ \code{nbNodes} a single \code{integer} corresponding
               to the number of nodes of the tested subnetwork. }
               \item{ \code{nbNodesOneLinkTested} a single \code{integer}
               corresponding to the number of nodes of the tested subnetwork
               when first-degree neighbor nodes are included. }
               \item{ \code{nbLinkOneLinkTested} a single \code{integer}
               corresponding to the number of links of the tested subnetwork
               including all links from the first-degree neighbor nodes.}
               \item{ \code{nbLinkTested} a single \code{integer}
               corresponding to the number of links of the tested subnetwork.}
               \item{ \code{seed} a single \code{numeric} corresponding
               to the seed used for the analysis.}
           }
}
\description{
Simulates a fixed number of subnetwork through randomness sampling.
Each subnetwork is containing the same fixed number of nodes. For all simulated
subetwork, three important informations are extracted: the number of
nodes present when including the first-degree neighbor nodes, the number
of links and the number of links when includes all links from the first-degree
neighbor nodes.
}
\examples{
## TODO


}
\author{
Pascal Belleau, Astrid Deschenes
}
\keyword{subnetwork}
